SDS Header pointer math rewritten in a more elegant form.
As suggested by unwind in the Hacker News site the calculation of "sh"
could be improved. In his own words:

"Since the entire idea is that the pointer on the left-hand side is to
the type whose size should be subtracted, I think it's better not to
repeat the type but to "lock it" to the pointer instead. This also (of
course) means we can drop the parenthesis with sizeof, since those are
only needed when its argument is a type name."