favor functions over macros
It looks like using functions instead of macros
we’re gaining some readability, as the macro required
several lines, and we’re enforcing the type checks.

I’ve added <stdbool.h> dependency, that if I’m not
mistaken should be create any issue as it’s defined
in C99 and library requires that C version.