Remove SDS_TYPE_5, use a clean macro lambda.
2 extra bytes is not a huge difference (it probably doesn't even make a
difference at all because of malloc alignment), and it is arguably more
efficient and safer.

This way, we have many advantages:
  - We don't have to worry about treating SDS_TYPE_5 differently
  - We don't have to worry about SDS_TYPE_5 not having allocation info
  - We don't have to reallocate as much; each modification of an
    SDS_TYPE_5's length **requires** a realloc call, ruining the purpose
    of overallocating.
  - We can reuse the same code for each SDS string type.
  - We only need two bits for the flags, so a bitwise and is guaranteed
    to cover all cases. This removes the need for a default case.

Because we can reuse the code, we can create a lambda to make the code
much clearer and less repetitive (of course, it does get expanded in
compile time, but we don't have to do it ourselves).

Syntax:

SDS_HDR_LAMBDA(sds s, block);
SDS_HDR_LAMBDA_2(sds s, unsigned char type, block);

Note: block is actually a __VA_ARGS__. so commas are preserved.

SDS_HDR_LAMBDA will automatically generate the switch statement, and
give you the following:
  - sh: appropriate sds header
  - sds_hdr_uint: typedef for the header's size unit, unsigned.
  - sds_hdr_int: typedef for the header's size unit, signed.

Instead of this:

    static inline size_t sdslen(const sds s) {
        unsigned char flags = s[-1];
        switch(flags&SDS_TYPE_MASK) {
            case SDS_TYPE_5:
                return SDS_TYPE_5_LEN(flags);
            case SDS_TYPE_8:
                return SDS_HDR(8,s)->len;
            case SDS_TYPE_16:
                return SDS_HDR(16,s)->len;
            case SDS_TYPE_32:
                return SDS_HDR(32,s)->len;
            case SDS_TYPE_64:
                return SDS_HDR(64,s)->len;
        }
        return 0;
    }

now, all that needs to be written is this:

    static inline size_t sdslen(const sds s) {
        SDS_HDR_LAMBDA(s, { return sh->len; });
        return 0;
    }

Note that we need a diagnostic pragma because GCC and Clang will
complain if you don't use a local typedef.