Introducing the magic sdsadd macro! (also tests, std::string, and sdssplit)
sdsadd is the new thing. It is sdsadd's += operator.

It is implemented in three ways, in order of preference:
 - _Generic (for C11): SDSADD_TYPE=1
 - C++ overloads/templates (for C++11): SDSADD_TYPE=2
 - GCC/clang extensions (for GCC/Clang, C mode only): SDSADD_TYPE=3

Support is automatically detected, with SDSADD_TYPE=0 disabling
sdsadd. However, if the macros are manually defined, sds will
try to use that implementation anyways.

Either way, sdsadd will let you add...
 - C strings
 - char literals*
 - integers (int, unsigned int, long long, unsigned long long)
 - std::strings (C++, obviously)

    sds s = sdsempty();
    s = sdsadd(s, "Hello");
    s = sdsadd(s, ' ');
    s = sdsadd(s, "world!");
    s = sdsadd(s, 1337);

It is not recommended to use sdsadd with sds strings, as it will be
treated as a normal C string, requiring a strlen call. Use sdscatsds
instead.

*Note: Because integer promotion messes up overloading and _Generic,
sdsadd will only detect character literals if they were previously
declared or cast as a char (not int) or if the statement starts or
ends with a single quote.

For the implementation, this is the SDS_IS_CHAR macro:

    #define SDS_IS_CHAR(x) ((#x[0] == '\'') || (sizeof(#x) > 3 && #x[sizeof(#x) - 2] == '\''))

That, as well as the normal char overload, matches most cases:

    char h = 'h';
    sdsadd(s, h)           -> ok
    sdsadd(s, 'c')         -> ok
    sdsadd(s, '0' + d)     -> ok
    sdsadd(s, d + '0')     -> ok
    sdsadd(s, (char)57)    -> ok

It does not match these:

    int c = 'c';
    sdsadd(s, c)           -> not ok
    sdsadd(s, ('0' + d))   -> not ok
    sdsadd(s, 1 + '0' + 1) -> not ok

    int8_t I = 'I'; (or uint8_t, signed char, unsigned char)
    sdsadd(s, I)           -> probably not
    sdsadd(s, (char)I)     -> ok

It does have one kind of false positive:

    sdsadd(s, (int)'h')    -> not ok

Just wrap it with parentheses or use a full sdsaddint to fix it.

    sdsaddint(s, (int)'h') -> ok
    sdsadd(s, (int)('h')); -> ok

I put way too much time into this.

Other small features:

There are now a lot more tests, as well as separate tests for base,
c11, gcc, and C++ implementations. By default, 'make' will build
all of them, but you can either run

    make base; ./sds-test
    make gcc; ./sds-test-gcc
    make c++; ./sds-test-c++
    make c11; ./sds-test-c11

to build and run the individual tests.

sds also adds basic support for converting to and from std::string.

    std::string sds2stdstr(sds s)
    sds sdsfromstdstr(const std::string &x)
    sds sdsaddstdstr(sds s, const std::string &x)

sdssplit is also added, it simply wraps sdssplitlen with strlen calls.