Fix 32-bit and likely support older GCC/all Clang versions
Presumably, this will now compile with GCC 3.2+ or any Clang
version, but I have to do more testing.

I also fixed switch/maybe uninitialized warnings on 32-bit.

I forgot to mention that I disabled SDS_TYPE_64 on 32-bit because it is
impossible for 32-bit to address that much memory, and it just becomes a
pointless check.