Fix a bunch of dumb errors, const sds,  clean up headers, test longer strings
A bunch of things are now fixed.
 - 32-bit works properly. `1 & 2 = 0`. Not 1.
 - MSVC now compiles.
 - sds functions should be `dllexport`ed now.
 - GCC extension sdsadd now works if an array has unknown size.
   - We need a strlen call, but that is minimal and probably
     able to be optimized out because sdscat is inline.
   - No more `#pragma GCC diagnostic` hacks.
   - Smaller macros.

In sds.c and sds.h only, `sds` is now a macro defined to `char *`.
Before this, `const sds` expanded to `char *const` instead of
`const char *`. Outside these files, it is back to a typedef.
This is useful for documenting that an sds is `const` without
breaking stuff like `sds x, y`.

sds.h now works with a C90 + long long compiler. C99/C++11 is still
required to compile sds.c.

Additionally, a bunch of macros are now out of sds.h and moved to
sds.c.

sds-test now tests a MUCH longer string of about 4090 characters.
This is to stress test reallocation and `sdscatprintf`.

If sds is compiled with `SDS_ABORT_ON_ERROR`, it will add
`__attribute__((__returns_nonnull__))` to return values in GCC for
potential optimizations.